[REWORK] make all requests async and fix several issues
In the previous implementation, all the requests were blocking.
Thus during push/pull, the user has to wait the action to be done
which can be long and annoying if we have large/many tasklists.

This patch change completely the initial implementation and make all
the requests async (except for tokens).

Here is an update of the actions usage:
- Push: push one or all taskslists
  -> once the push is done, we pull the tasklists pushed

- Pull: pull one or all taskslists

- Add: add taskslists which are not present on the gtasks account
  -> when a new org file is created in the corresponding directory,
     it's detected and this file (tasklist) can be added.

- Remove: remove taskslists which are present on the gtasks account
  -> when a org file is removed in the corresponding directory,
     if this file is present in tasklists, it can be removed.

This patch contains also several fixes:
- "correctly" re-sync after push
- push only org element level 1 as task
- add "completed" only when it's needed
- extract/clean notes
- better errors code handling

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>